Added lap handling the arduino Changed timing from float to double

+++ server/src/main/java/com/antigravity/protocols/ProtocolListener.java
@@ -1,5 +1,5 @@
 package com.antigravity.protocols;
 
 public interface ProtocolListener {
-    void onLap(int lane, float lapTime);
+    void onLap(int lane, double lapTime);
 }
